Moved to Eclipse

+++ Logisim-Fork/com/cburch/logisim/data/Value.java
@@ -9,11 +9,11 @@ import java.util.Arrays;
 import com.cburch.logisim.util.Cache;
 
 public class Value {
-	public static final Value FALSE   = new Value(1, 0, 0, 0);
-	public static final Value TRUE    = new Value(1, 0, 0, 1);
+	public static final Value FALSE = new Value(1, 0, 0, 0);
+	public static final Value TRUE = new Value(1, 0, 0, 1);
 	public static final Value UNKNOWN = new Value(1, 0, 1, 0);
-	public static final Value ERROR   = new Value(1, 1, 0, 0);
-	public static final Value NIL     = new Value(0, 0, 0, 0);
+	public static final Value ERROR = new Value(1, 1, 0, 0);
+	public static final Value NIL = new Value(0, 0, 0, 0);
 
 	public static final int MAX_WIDTH = 32;
 
@@ -24,14 +24,16 @@ public class Value {
 	public static final Color ERROR_COLOR = new Color(192, 0, 0);
 	public static final Color WIDTH_ERROR_COLOR = new Color(255, 123, 0);
 	public static final Color MULTI_COLOR = Color.BLACK;
-	
+
 	private static final Cache cache = new Cache();
 
 	public static Value create(Value[] values) {
-		if (values.length == 0) return NIL;
-		if (values.length == 1) return values[0];
-		if (values.length > MAX_WIDTH) throw new RuntimeException(
-			"Cannot have more than " + MAX_WIDTH + " bits in a value");
+		if (values.length == 0)
+			return NIL;
+		if (values.length == 1)
+			return values[0];
+		if (values.length > MAX_WIDTH)
+			throw new RuntimeException("Cannot have more than " + MAX_WIDTH + " bits in a value");
 
 		int width = values.length;
 		int value = 0;
@@ -39,13 +41,15 @@ public class Value {
 		int error = 0;
 		for (int i = 0; i < values.length; i++) {
 			int mask = 1 << i;
-			if (values[i] == TRUE)         value |= mask;
-			else if (values[i] == FALSE)   /* do nothing */;
-			else if (values[i] == UNKNOWN) unknown |= mask;
-			else if (values[i] == ERROR)   error |= mask;
+			if (values[i] == TRUE)
+				value |= mask;
+			else if (values[i] == FALSE) /* do nothing */;
+			else if (values[i] == UNKNOWN)
+				unknown |= mask;
+			else if (values[i] == ERROR)
+				error |= mask;
 			else {
-				throw new RuntimeException("unrecognized value "
-					+ values[i]);
+				throw new RuntimeException("unrecognized value " + values[i]);
 			}
 		}
 		return Value.create(width, error, unknown, value);
@@ -67,10 +71,14 @@ public class Value {
 		if (width == 0) {
 			return Value.NIL;
 		} else if (width == 1) {
-			if ((error & 1) != 0)       return Value.ERROR;
-			else if ((unknown & 1) != 0)    return Value.UNKNOWN;
-			else if ((value & 1) != 0)  return Value.TRUE;
-			else return Value.FALSE;
+			if ((error & 1) != 0)
+				return Value.ERROR;
+			else if ((unknown & 1) != 0)
+				return Value.UNKNOWN;
+			else if ((value & 1) != 0)
+				return Value.TRUE;
+			else
+				return Value.FALSE;
 		} else {
 			int mask = (width == 32 ? -1 : ~(-1 << width));
 			error = error & mask;
@@ -81,15 +89,15 @@ public class Value {
 			Object cached = cache.get(hashCode);
 			if (cached != null) {
 				Value val = (Value) cached;
-				if (val.value == value && val.width == width && val.error == error
-						&& val.unknown == unknown) return val;
+				if (val.value == value && val.width == width && val.error == error && val.unknown == unknown)
+					return val;
 			}
-			Value ret= new Value(width, error, unknown, value);
+			Value ret = new Value(width, error, unknown, value);
 			cache.put(hashCode, ret);
 			return ret;
 		}
 	}
-	
+
 	public static Value repeat(Value base, int bits) {
 		if (base.getWidth() != 1) {
 			throw new IllegalArgumentException("first parameter must be one bit");
@@ -109,7 +117,8 @@ public class Value {
 	private final int value;
 
 	private Value(int width, int error, int unknown, int value) {
-		// To ensure that the one-bit values are unique, this should be called only
+		// To ensure that the one-bit values are unique, this should be called
+		// only
 		// for the one-bit values and by the private create method
 		this.width = width;
 		this.error = error;
@@ -122,7 +131,8 @@ public class Value {
 	}
 
 	public Value extendWidth(int newWidth, Value others) {
-		if (width == newWidth) return this;
+		if (width == newWidth)
+			return this;
 		int maskInverse = (width == 32 ? 0 : (-1 << width));
 		if (others == Value.ERROR) {
 			return Value.create(newWidth, error | maskInverse, unknown, value);
@@ -156,10 +166,8 @@ public class Value {
 			return val;
 		} else {
 			int mask = ~(1 << which);
-			return Value.create(this.width,
-				(this.error   & mask) | (val.error   << which),
-				(this.unknown & mask) | (val.unknown << which),
-				(this.value   & mask) | (val.value   << which));
+			return Value.create(this.width, (this.error & mask) | (val.error << which),
+					(this.unknown & mask) | (val.unknown << which), (this.value & mask) | (val.value << which));
 		}
 	}
 
@@ -172,12 +180,17 @@ public class Value {
 	}
 
 	public Value get(int which) {
-		if (which < 0 || which >= width) return ERROR;
+		if (which < 0 || which >= width)
+			return ERROR;
 		int mask = 1 << which;
-		if ((error & mask) != 0) return ERROR;
-		else if ((unknown & mask) != 0) return UNKNOWN;
-		else if ((value & mask) != 0) return TRUE;
-		else return FALSE;
+		if ((error & mask) != 0)
+			return ERROR;
+		else if ((unknown & mask) != 0)
+			return UNKNOWN;
+		else if ((value & mask) != 0)
+			return TRUE;
+		else
+			return FALSE;
 	}
 
 	public BitWidth getBitWidth() {
@@ -190,15 +203,14 @@ public class Value {
 
 	@Override
 	public boolean equals(Object other_obj) {
-		if (!(other_obj instanceof Value)) return false;
+		if (!(other_obj instanceof Value))
+			return false;
 		Value other = (Value) other_obj;
-		boolean ret = this.width == other.width
-			&& this.error == other.error
-			&& this.unknown == other.unknown
-			&& this.value == other.value;
+		boolean ret = this.width == other.width && this.error == other.error && this.unknown == other.unknown
+				&& this.value == other.value;
 		return ret;
 	}
-	
+
 	@Override
 	public int hashCode() {
 		int ret = width;
@@ -209,30 +221,38 @@ public class Value {
 	}
 
 	public int toIntValue() {
-		if (error != 0) return -1;
-		if (unknown != 0) return -1;
+		if (error != 0)
+			return -1;
+		if (unknown != 0)
+			return -1;
 		return value;
 	}
 
 	@Override
 	public String toString() {
 		switch (width) {
-		case 0: return "-";
+		case 0:
+			return "-";
 		case 1:
-			if (error != 0)        return "E";
-			else if (unknown != 0) return "x";
-			else if (value != 0)   return "1";
-			else                  return "0";
+			if (error != 0)
+				return "E";
+			else if (unknown != 0)
+				return "x";
+			else if (value != 0)
+				return "1";
+			else
+				return "0";
 		default:
 			StringBuilder ret = new StringBuilder();
 			for (int i = width - 1; i >= 0; i--) {
 				ret.append(get(i).toString());
-				if (i % 4 == 0 && i != 0) ret.append(" ");
+				if (i % 4 == 0 && i != 0)
+					ret.append(" ");
 			}
 			return ret.toString();
 		}
 	}
-	
+
 	public String toOctalString() {
 		if (width <= 1) {
 			return toString();
@@ -246,17 +266,25 @@ public class Value {
 				int v = 0;
 				c[i] = '?';
 				for (int j = last - 1; j >= frst; j--) {
-					if (vals[j] == Value.ERROR) { c[i] = 'E'; break; }
-					if (vals[j] == Value.UNKNOWN) { c[i] = 'x'; break; }
+					if (vals[j] == Value.ERROR) {
+						c[i] = 'E';
+						break;
+					}
+					if (vals[j] == Value.UNKNOWN) {
+						c[i] = 'x';
+						break;
+					}
 					v = 2 * v;
-					if (vals[j] == Value.TRUE) v++;
+					if (vals[j] == Value.TRUE)
+						v++;
 				}
-				if (c[i] == '?') c[i] = Character.forDigit(v, 8);
+				if (c[i] == '?')
+					c[i] = Character.forDigit(v, 8);
 			}
 			return new String(c);
 		}
 	}
-	
+
 	public String toHexString() {
 		if (width <= 1) {
 			return toString();
@@ -270,22 +298,33 @@ public class Value {
 				int v = 0;
 				c[i] = '?';
 				for (int j = last - 1; j >= frst; j--) {
-					if (vals[j] == Value.ERROR) { c[i] = 'E'; break; }
-					if (vals[j] == Value.UNKNOWN) { c[i] = 'x'; break; }
+					if (vals[j] == Value.ERROR) {
+						c[i] = 'E';
+						break;
+					}
+					if (vals[j] == Value.UNKNOWN) {
+						c[i] = 'x';
+						break;
+					}
 					v = 2 * v;
-					if (vals[j] == Value.TRUE) v++;
+					if (vals[j] == Value.TRUE)
+						v++;
 				}
-				if (c[i] == '?') c[i] = Character.forDigit(v, 16);
+				if (c[i] == '?')
+					c[i] = Character.forDigit(v, 16);
 			}
 			return new String(c);
 		}
 	}
-	
+
 	public String toDecimalString(boolean signed) {
-		if (width == 0) return "-";
-		if (isErrorValue()) return Strings.get("valueError");
-		if (!isFullyDefined()) return Strings.get("valueUnknown");
-		
+		if (width == 0)
+			return "-";
+		if (isErrorValue())
+			return Strings.get("valueError");
+		if (!isFullyDefined())
+			return Strings.get("valueUnknown");
+
 		int value = toIntValue();
 		if (signed) {
 			if (width < 32 && (value >> (width - 1)) != 0) {
@@ -293,120 +332,139 @@ public class Value {
 			}
 			return "" + value;
 		} else {
-			return "" + ((long) value & 0xFFFFFFFFL);
+			return "" + (value & 0xFFFFFFFFL);
 		}
 	}
-	
+
 	public String toDisplayString(int radix) {
 		switch (radix) {
-		case 2:  return toDisplayString();
-		case 8:  return toOctalString();
-		case 16: return toHexString();
+		case 2:
+			return toDisplayString();
+		case 8:
+			return toOctalString();
+		case 16:
+			return toHexString();
 		default:
-			if (width == 0) return "-";
-			if (isErrorValue()) return Strings.get("valueError");
-			if (!isFullyDefined()) return Strings.get("valueUnknown");
+			if (width == 0)
+				return "-";
+			if (isErrorValue())
+				return Strings.get("valueError");
+			if (!isFullyDefined())
+				return Strings.get("valueUnknown");
 			return Integer.toString(toIntValue(), radix);
 		}
 	}
 
 	public String toDisplayString() {
 		switch (width) {
-		case 0: return "-";
+		case 0:
+			return "-";
 		case 1:
-			if (error != 0)        return Strings.get("valueErrorSymbol");
-			else if (unknown != 0) return Strings.get("valueUnknownSymbol");
-			else if (value != 0)   return "1";
-			else                  return "0";
+			if (error != 0)
+				return Strings.get("valueErrorSymbol");
+			else if (unknown != 0)
+				return Strings.get("valueUnknownSymbol");
+			else if (value != 0)
+				return "1";
+			else
+				return "0";
 		default:
 			StringBuilder ret = new StringBuilder();
 			for (int i = width - 1; i >= 0; i--) {
 				ret.append(get(i).toString());
-				if (i % 4 == 0 && i != 0) ret.append(" ");
+				if (i % 4 == 0 && i != 0)
+					ret.append(" ");
 			}
 			return ret.toString();
 		}
 	}
 
 	public Value combine(Value other) {
-		if (other == null) return this;
-		if (this == NIL) return other;
-		if (other == NIL) return this;
+		if (other == null)
+			return this;
+		if (this == NIL)
+			return other;
+		if (other == NIL)
+			return this;
 		if (this.width == 1 && other.width == 1) {
-			if (this == other) return this;
-			if (this == UNKNOWN) return other;
-			if (other == UNKNOWN) return this;
+			if (this == other)
+				return this;
+			if (this == UNKNOWN)
+				return other;
+			if (other == UNKNOWN)
+				return this;
 			return ERROR;
 		} else {
-			int disagree = (this.value ^ other.value)
-				& ~(this.unknown | other.unknown);
-			return Value.create(Math.max(this.width, other.width),
-				this.error | other.error | disagree,
-				this.unknown & other.unknown,
-				(this.value & ~this.unknown) | (other.value & ~other.unknown));
+			int disagree = (this.value ^ other.value) & ~(this.unknown | other.unknown);
+			return Value.create(Math.max(this.width, other.width), this.error | other.error | disagree,
+					this.unknown & other.unknown, (this.value & ~this.unknown) | (other.value & ~other.unknown));
 		}
 	}
 
 	public Value and(Value other) {
-		if (other == null) return this;
+		if (other == null)
+			return this;
 		if (this.width == 1 && other.width == 1) {
-			if (this == FALSE || other == FALSE) return FALSE;
-			if (this == TRUE  && other == TRUE ) return TRUE;
+			if (this == FALSE || other == FALSE)
+				return FALSE;
+			if (this == TRUE && other == TRUE)
+				return TRUE;
 			return ERROR;
 		} else {
 			int false0 = ~this.value & ~this.error & ~this.unknown;
 			int false1 = ~other.value & ~other.error & ~other.unknown;
 			int falses = false0 | false1;
 			return Value.create(Math.max(this.width, other.width),
-					(this.error | other.error | this.unknown | other.unknown) & ~falses,
-					0,
-					this.value & other.value);
+					(this.error | other.error | this.unknown | other.unknown) & ~falses, 0, this.value & other.value);
 		}
 	}
 
 	public Value or(Value other) {
-		if (other == null) return this;
+		if (other == null)
+			return this;
 		if (this.width == 1 && other.width == 1) {
-			if (this == TRUE  || other == TRUE ) return TRUE;
-			if (this == FALSE && other == FALSE) return FALSE;
+			if (this == TRUE || other == TRUE)
+				return TRUE;
+			if (this == FALSE && other == FALSE)
+				return FALSE;
 			return ERROR;
 		} else {
 			int true0 = this.value & ~this.error & ~this.unknown;
 			int true1 = other.value & ~other.error & ~other.unknown;
 			int trues = true0 | true1;
 			return Value.create(Math.max(this.width, other.width),
-				(this.error | other.error | this.unknown | other.unknown) & ~trues,
-				0,
-				this.value | other.value);
+					(this.error | other.error | this.unknown | other.unknown) & ~trues, 0, this.value | other.value);
 		}
 	}
 
 	public Value xor(Value other) {
-		if (other == null) return this;
+		if (other == null)
+			return this;
 		if (this.width <= 1 && other.width <= 1) {
-			if (this == ERROR || other == ERROR) return ERROR;
-			if (this == UNKNOWN || other == UNKNOWN) return ERROR;
-			if (this == NIL || other == NIL) return ERROR;
-			if ((this == TRUE) == (other == TRUE)) return FALSE;
+			if (this == ERROR || other == ERROR)
+				return ERROR;
+			if (this == UNKNOWN || other == UNKNOWN)
+				return ERROR;
+			if (this == NIL || other == NIL)
+				return ERROR;
+			if ((this == TRUE) == (other == TRUE))
+				return FALSE;
 			return TRUE;
 		} else {
 			return Value.create(Math.max(this.width, other.width),
-				this.error | other.error | this.unknown | other.unknown,
-				0,
-				this.value ^ other.value);
+					this.error | other.error | this.unknown | other.unknown, 0, this.value ^ other.value);
 		}
 	}
 
 	public Value not() {
 		if (width <= 1) {
-			if (this == TRUE) return FALSE;
-			if (this == FALSE) return TRUE;
+			if (this == TRUE)
+				return FALSE;
+			if (this == FALSE)
+				return TRUE;
 			return ERROR;
 		} else {
-			return Value.create(this.width,
-				this.error | this.unknown,
-				0,
-				~this.value);
+			return Value.create(this.width, this.error | this.unknown, 0, ~this.value);
 		}
 	}
 
@@ -416,9 +474,12 @@ public class Value {
 		} else if (width == 0) {
 			return NIL_COLOR;
 		} else if (width == 1) {
-			if (this == UNKNOWN) return UNKNOWN_COLOR;
-			else if (this == TRUE) return TRUE_COLOR;
-			else return FALSE_COLOR;
+			if (this == UNKNOWN)
+				return UNKNOWN_COLOR;
+			else if (this == TRUE)
+				return TRUE_COLOR;
+			else
+				return FALSE_COLOR;
 		} else {
 			return MULTI_COLOR;
 		}
