+++ Logisim-Fork/com/cburch/logisim/std/arith/Divider.java
@@ -23,26 +23,25 @@ import com.cburch.logisim.util.GraphicsUtil;
 public class Divider extends InstanceFactory {
 	static final int PER_DELAY = 1;
 
-	private static final int IN0   = 0;
-	private static final int IN1   = 1;
-	private static final int OUT   = 2;
+	private static final int IN0 = 0;
+	private static final int IN1 = 1;
+	private static final int OUT = 2;
 	private static final int UPPER = 3;
-	private static final int REM   = 4;
+	private static final int REM = 4;
 
 	public Divider() {
 		super("Divider", Strings.getter("dividerComponent"));
-		setAttributes(new Attribute[] { StdAttr.WIDTH },
-				new Object[] { BitWidth.create(8) });
+		setAttributes(new Attribute[] { StdAttr.WIDTH }, new Object[] { BitWidth.create(8) });
 		setKeyConfigurator(new BitWidthConfigurator(StdAttr.WIDTH));
 		setOffsetBounds(Bounds.create(-40, -20, 40, 40));
 		setIconName("divider.gif");
-		
+
 		Port[] ps = new Port[5];
-		ps[IN0]   = new Port(-40, -10, Port.INPUT,  StdAttr.WIDTH);
-		ps[IN1]   = new Port(-40,  10, Port.INPUT,  StdAttr.WIDTH);
-		ps[OUT]   = new Port(  0,   0, Port.OUTPUT, StdAttr.WIDTH);
-		ps[UPPER] = new Port(-20, -20, Port.INPUT,  StdAttr.WIDTH);
-		ps[REM]   = new Port(-20,  20, Port.OUTPUT, StdAttr.WIDTH);
+		ps[IN0] = new Port(-40, -10, Port.INPUT, StdAttr.WIDTH);
+		ps[IN1] = new Port(-40, 10, Port.INPUT, StdAttr.WIDTH);
+		ps[OUT] = new Port(0, 0, Port.OUTPUT, StdAttr.WIDTH);
+		ps[UPPER] = new Port(-20, -20, Port.INPUT, StdAttr.WIDTH);
+		ps[REM] = new Port(-20, 20, Port.OUTPUT, StdAttr.WIDTH);
 		ps[IN0].setToolTip(Strings.getter("dividerDividendLowerTip"));
 		ps[IN1].setToolTip(Strings.getter("dividerDivisorTip"));
 		ps[OUT].setToolTip(Strings.getter("dividerOutputTip"));
@@ -77,7 +76,7 @@ public class Divider extends InstanceFactory {
 		painter.drawPort(IN0);
 		painter.drawPort(IN1);
 		painter.drawPort(OUT);
-		painter.drawPort(UPPER, Strings.get("dividerUpperInput"),  Direction.NORTH);
+		painter.drawPort(UPPER, Strings.get("dividerUpperInput"), Direction.NORTH);
 		painter.drawPort(REM, Strings.get("dividerRemainderOutput"), Direction.SOUTH);
 
 		Location loc = painter.getLocation();
@@ -93,12 +92,13 @@ public class Divider extends InstanceFactory {
 
 	static Value[] computeResult(BitWidth width, Value a, Value b, Value upper) {
 		int w = width.getWidth();
-		if (upper == Value.NIL || upper.isUnknown()) upper = Value.createKnown(width, 0);
+		if (upper == Value.NIL || upper.isUnknown())
+			upper = Value.createKnown(width, 0);
 		if (a.isFullyDefined() && b.isFullyDefined() && upper.isFullyDefined()) {
-			long num = ((long) upper.toIntValue() << w)
-				| ((long) a.toIntValue() & 0xFFFFFFFFL);
-			long den = (long) b.toIntValue() & 0xFFFFFFFFL;
-			if (den == 0) den = 1;
+			long num = ((long) upper.toIntValue() << w) | (a.toIntValue() & 0xFFFFFFFFL);
+			long den = b.toIntValue() & 0xFFFFFFFFL;
+			if (den == 0)
+				den = 1;
 			long result = num / den;
 			long rem = num % den;
 			if (rem < 0) {
@@ -110,8 +110,7 @@ public class Divider extends InstanceFactory {
 					result++;
 				}
 			}
-			return new Value[] { Value.createKnown(width, (int) result),
-					Value.createKnown(width, (int) rem) };
+			return new Value[] { Value.createKnown(width, (int) result), Value.createKnown(width, (int) rem) };
 		} else if (a.isErrorValue() || b.isErrorValue() || upper.isErrorValue()) {
 			return new Value[] { Value.createError(width), Value.createError(width) };
 		} else {
