+++ Logisim-Fork/com/cburch/logisim/std/arith/Multiplier.java
@@ -23,26 +23,25 @@ import com.cburch.logisim.util.GraphicsUtil;
 public class Multiplier extends InstanceFactory {
 	static final int PER_DELAY = 1;
 
-	private static final int IN0   = 0;
-	private static final int IN1   = 1;
-	private static final int OUT   = 2;
-	private static final int C_IN  = 3;
+	private static final int IN0 = 0;
+	private static final int IN1 = 1;
+	private static final int OUT = 2;
+	private static final int C_IN = 3;
 	private static final int C_OUT = 4;
 
 	public Multiplier() {
 		super("Multiplier", Strings.getter("multiplierComponent"));
-		setAttributes(new Attribute[] { StdAttr.WIDTH },
-				new Object[] { BitWidth.create(8) });
+		setAttributes(new Attribute[] { StdAttr.WIDTH }, new Object[] { BitWidth.create(8) });
 		setKeyConfigurator(new BitWidthConfigurator(StdAttr.WIDTH));
 		setOffsetBounds(Bounds.create(-40, -20, 40, 40));
 		setIconName("multiplier.gif");
 
 		Port[] ps = new Port[5];
-		ps[IN0]   = new Port(-40, -10, Port.INPUT,  StdAttr.WIDTH);
-		ps[IN1]   = new Port(-40,  10, Port.INPUT,  StdAttr.WIDTH);
-		ps[OUT]   = new Port(  0,   0, Port.OUTPUT, StdAttr.WIDTH);
-		ps[C_IN]  = new Port(-20, -20, Port.INPUT,  StdAttr.WIDTH);
-		ps[C_OUT] = new Port(-20,  20, Port.OUTPUT, StdAttr.WIDTH);
+		ps[IN0] = new Port(-40, -10, Port.INPUT, StdAttr.WIDTH);
+		ps[IN1] = new Port(-40, 10, Port.INPUT, StdAttr.WIDTH);
+		ps[OUT] = new Port(0, 0, Port.OUTPUT, StdAttr.WIDTH);
+		ps[C_IN] = new Port(-20, -20, Port.INPUT, StdAttr.WIDTH);
+		ps[C_OUT] = new Port(-20, 20, Port.OUTPUT, StdAttr.WIDTH);
 		ps[IN0].setToolTip(Strings.getter("multiplierInputTip"));
 		ps[IN1].setToolTip(Strings.getter("multiplierInputTip"));
 		ps[OUT].setToolTip(Strings.getter("multiplierOutputTip"));
@@ -64,7 +63,7 @@ public class Multiplier extends InstanceFactory {
 
 		// propagate them
 		int delay = dataWidth.getWidth() * (dataWidth.getWidth() + 2) * PER_DELAY;
-		state.setPort(OUT,   outs[0], delay);
+		state.setPort(OUT, outs[0], delay);
 		state.setPort(C_OUT, outs[1], delay);
 	}
 
@@ -77,7 +76,7 @@ public class Multiplier extends InstanceFactory {
 		painter.drawPort(IN0);
 		painter.drawPort(IN1);
 		painter.drawPort(OUT);
-		painter.drawPort(C_IN,  "c in",  Direction.NORTH);
+		painter.drawPort(C_IN, "c in", Direction.NORTH);
 		painter.drawPort(C_OUT, "c out", Direction.SOUTH);
 
 		Location loc = painter.getLocation();
@@ -90,15 +89,13 @@ public class Multiplier extends InstanceFactory {
 		GraphicsUtil.switchToWidth(g, 1);
 	}
 
-
 	static Value[] computeProduct(BitWidth width, Value a, Value b, Value c_in) {
 		int w = width.getWidth();
-		if (c_in == Value.NIL || c_in.isUnknown()) c_in = Value.createKnown(width, 0);
+		if (c_in == Value.NIL || c_in.isUnknown())
+			c_in = Value.createKnown(width, 0);
 		if (a.isFullyDefined() && b.isFullyDefined() && c_in.isFullyDefined()) {
-			long sum = (long) a.toIntValue() * (long) b.toIntValue()
-				+ (long) c_in.toIntValue();
-			return new Value[] { Value.createKnown(width, (int) sum),
-				Value.createKnown(width, (int) (sum >> w)) };
+			long sum = (long) a.toIntValue() * (long) b.toIntValue() + c_in.toIntValue();
+			return new Value[] { Value.createKnown(width, (int) sum), Value.createKnown(width, (int) (sum >> w)) };
 		} else {
 			Value[] avals = a.getAll();
 			int aOk = findUnknown(avals);
@@ -112,7 +109,7 @@ public class Multiplier extends InstanceFactory {
 			int cOk = findUnknown(cvals);
 			int cErr = findError(cvals);
 			int cx = getKnown(cvals);
-			
+
 			int known = Math.min(Math.min(aOk, bOk), cOk);
 			int error = Math.min(Math.min(aErr, bErr), cErr);
 			int ret = ax * bx + cx;
@@ -131,26 +128,29 @@ public class Multiplier extends InstanceFactory {
 					error < w ? Value.createError(width) : Value.createUnknown(width) };
 		}
 	}
-	
+
 	private static int findUnknown(Value[] vals) {
 		for (int i = 0; i < vals.length; i++) {
-			if (!vals[i].isFullyDefined()) return i;
+			if (!vals[i].isFullyDefined())
+				return i;
 		}
 		return vals.length;
 	}
-	
+
 	private static int findError(Value[] vals) {
 		for (int i = 0; i < vals.length; i++) {
-			if (vals[i].isErrorValue()) return i;
+			if (vals[i].isErrorValue())
+				return i;
 		}
 		return vals.length;
 	}
-	
+
 	private static int getKnown(Value[] vals) {
 		int ret = 0;
 		for (int i = 0; i < vals.length; i++) {
 			int val = vals[i].toIntValue();
-			if (val < 0) return ret;
+			if (val < 0)
+				return ret;
 			ret |= val << i;
 		}
 		return ret;
